requirements.txt
fastapi==0.111.0
openai==0.28.0
pdfplumber==0.11.0
python-dotenv==0.20.0
python-multipart==0.0.9
streamlit==1.34.0
uvicorn==0.29.0
